requirements/base.txt
Django==1.11.6
argparse==1.4.0

djangorestframework==3.5.3
django-coreapi==2.3.0
olefile==0.44

# Images
Pillow==3.4.2

# Custom user with email
django-authtools==1.6.0

# Postgres
psycopg2==2.7.1

# Utils
requests==2.18.1
django-extensions==1.8.1
ipdb==0.10.3

# Db Backups
# Upgrade whenever a fix comes out https://github.com/django-dbbackup/django-dbbackup/issues/221
# django-dbbackup==3.0.2
# Waiting for a pypi release.
git+https://github.com/django-dbbackup/django-dbbackup.git@5064a92187e540134a0d9b0cf8d1dd533a3c10a6

# Testing
coverage==4.3.4

# Redis cache backend
redis==2.10.5
django-redis-cache==1.7.1

# Sentry
raven==6.2.1
